requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
groq>=0.11.0
python-dotenv==1.0.0
openai==1.12.0
Pillow>=10.0.0
requests==2.31.0